Splitting keyring into its own module so that we can load without this as a dependency
# Conflicts:
#	msticpy/common/secret_settings.py
